+++ examples/mezzanine_example/runserver-bootstrap3-theme-mezzanine-integration.sh
@@ -1,0 +1,2 @@
+#workon fobi
+./manage.py runserver 0.0.0.0:8003 --settings=settings_bootstrap3_theme_mezzanine --traceback -v 3