+++ examples/mezzanine_example/settings_bootstrap3_theme_mezzanine.py
@@ -1,0 +1,4 @@
+from settings import *
+
+# Fobi extra configuration
+FOBI_DEFAULT_THEME = 'bootstrap3'