+++ examples/simple/context_processors.py
@@ -1,0 +1,5 @@
+def disable_admin_tools(request):
+    """
+    Disable admin tools.
+    """
+    return {'ADMIN_TOOLS_DISABLED': True}