+++ src/fobi/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'django-fobi'
-__version__ = '0.4.4'
-__build__ = 0x000013
+__version__ = '0.4.5'
+__build__ = 0x000014
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'