+++ src/fobi/contrib/apps/mezzanine_integration/defaults.py
@@ -1,0 +1,14 @@
+__title__ = 'fobi.contrib.apps.mezzanine_integration.defaults'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = (
+    'WIDGET_FORM_SENT_GET_PARAM', 'FORM_TEMPLATE_CHOICES',
+    'SUCCESS_PAGE_TEMPLATE_CHOICES',
+)
+
+WIDGET_FORM_SENT_GET_PARAM = 'sent'
+
+FORM_TEMPLATE_CHOICES = []
+
+SUCCESS_PAGE_TEMPLATE_CHOICES = []